+++ CHANGELOG.md
@@ -2,6 +2,26 @@
 
 All notable changes to this project will be documented in this file.
 
+## [1.2.9] - 2025-10-27
+
+### Introduced
+- Suppressing finding will make finding suppressed across new branches
+
+### Fixed
+- Suppress now working in both coderepo and team context
+
+## [1.2.8] - 2025-10-21
+
+### Introduced
+- Possibility to export results to XLSX
+- Possibility to add comment to suppress rule
+
+### Fixed
+- ScanInfo created when limit is met
+- Scan Scheduler that run every 8 hours
+- Handing of scans status
+
+
 ## [1.2.7] - 2025-09-18
 
 ### Introduced